+++ R/wc-inspect.R
@@ -5,16 +5,24 @@
 #' @md
 #' @param url URL to fetch
 #' @param js_delay (ms) How long to wait for JavaScript to execute/XHRs to load? (Default: 5000)
-#' @param timeout Sets the timeout (milliseconds) of the webc onnection. Set to zero for an infinite wait.
+#' @param timeout Sets the timeout (milliseconds) of the web connection. Set to zero for an infinite wait.
 #'        Defaults to `30000`. Note: The timeout is used twice. The first is for making the socket
 #'        connection, the second is for data retrieval. If the time is critical you must allow for twice
 #'        the time specified here.
+#' @param css,images enable CSS/download images? (default `FALSE`)
 #' @export
-wc_inspect <- function(url, js_delay = 5000L, timeout = 30000L) {
+wc_inspect <- function(url, js_delay = 5000L, timeout = 30000L, css = FALSE, images = FALSE) {
 
   app <- J("is.rud.htmlunit.Zapp")
 
-  res <- app$getRequestsFor(url, .jlong(js_delay), .jint(timeout))
+  app$getRequestsFor(
+    url,
+    .jlong(js_delay),
+    as.integer(timeout),
+    .jnew("java/lang/Boolean", css),
+    .jnew("java/lang/Boolean", images)
+  ) -> res
+
   res <- as.list(res)
 
   lapply(res, function(.x) {