+++ java/htmlunit/src/main/java/is/rud/htmlunit/RDefaultCssErrorHandler.java
@@ -1,0 +1,17 @@
+package is.rud.htmlunit;
+
+public class RDefaultCssErrorHandler implements com.gargoylesoftware.css.parser.CSSErrorHandler,
+                                                java.io.Serializable {
+    @Override
+    public void error(final com.gargoylesoftware.css.parser.CSSParseException exception) {
+    }
+
+    @Override
+    public void fatalError(final com.gargoylesoftware.css.parser.CSSParseException exception) {
+    }
+
+    @Override
+    public void warning(final com.gargoylesoftware.css.parser.CSSParseException exception) {
+    }
+
+}