+++ java/htmlunit/src/main/java/is/rud/htmlunit/RIncorrectnessListener.java
@@ -1,0 +1,13 @@
+package is.rud.htmlunit;
+
+public class RIncorrectnessListener implements com.gargoylesoftware.htmlunit.IncorrectnessListener,
+                                               java.io.Serializable {
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void notify(final java.lang.String message, final java.lang.Object origin) {
+    }
+
+}