+++ java/htmlunit/src/main/java/is/rud/htmlunit/Zapp.java
@@ -2,20 +2,28 @@ package is.rud.htmlunit;
 
 import com.gargoylesoftware.htmlunit.*;
 import com.gargoylesoftware.htmlunit.util.*;
+
 import java.util.*;
 import java.lang.*;
 import java.io.*;
 
 public class Zapp {
 
-  public static List<WebResponse> getRequestsFor(String url, long jsDelay, int timeout) throws IOException {
+  private static com.gargoylesoftware.htmlunit.IncorrectnessListener incorrectnessListener_ = new RIncorrectnessListener();
+  private static com.gargoylesoftware.css.parser.CSSErrorHandler cssErrorHandler_ = new RDefaultCssErrorHandler();
+
+  public static List<WebResponse> getRequestsFor(String url, long jsDelay, int timeout, Boolean css, Boolean images) throws IOException {
 
     final WebClient webClient = new WebClient(BrowserVersion.CHROME);
 
+    webClient.setCssErrorHandler(cssErrorHandler_);
+    webClient.setIncorrectnessListener(incorrectnessListener_);
+
     WebClientOptions wco = webClient.getOptions();
+
     wco.setThrowExceptionOnScriptError(false);
-    wco.setCssEnabled(true);
-    wco.setDownloadImages(true);
+    wco.setCssEnabled(css);
+    wco.setDownloadImages(images);
     wco.setTimeout(timeout);
 
     final List<WebResponse> list = new ArrayList<>();