+++ NEWS.md
@@ -1,6 +1,7 @@
 0.5.0
 * Updated for 2.43.0 jars
-* Added `timeout` to `wc_inspect()`
+* Added support for Microsoft Edge browser
+* Added `timeout`, `css`, and `images` parameters to `wc_inspect()`
 
 0.4.0
 * Switched to {tinytest}